requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
pydantic